feature(exec) : configuration d'un executable
