+++ .vscode/settings.json
@@ -7,6 +7,7 @@
         "dépubliée",
         "dépubliées",
         "dépublier",
+        "egpf",
         "EPSG",
         "fontawesome",
         "Géoplateforme",
